moved tab-controller/Tab to tab/index
